% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_assembly.R
\name{cov_apply}
\alias{cov_apply}
\title{Apply covariates to PK(PD) dataset}
\usage{
cov_apply(
  df,
  cov,
  id.by = "USUBJID",
  time.by = NA,
  direction = "downup",
  cov.rnd = NA,
  na = -999
)
}
\arguments{
\item{df}{PK(PD) dataframe generated by pk_build}

\item{cov}{dataframe of covariates}

\item{id.by}{id variable to merge covariates}

\item{time.by}{time variable to merge covariates}

\item{direction}{fill direction for time-varying covariates}

\item{cov.rnd}{covariate rounding parameter}

\item{na}{value to replace NA numeric covariates}
}
\value{
PK(PD) dataset with additional covariates
}
\description{
Apply covariates to PK(PD) dataset
}
