% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\docType{data}
\name{DM}
\alias{DM}
\title{DM}
\format{
## `DM`
A data frame with 22 rows and 12 variables:
\describe{
  \item{STUDYID}{study label}
  \item{SITEID}{site code}
  \item{SUBJID}{subject code}
  \item{USUBJID}{unique subject ID}
  \item{SCRFL}{screen fail flag}
  \item{ICFDT}{informed consent date}
  \item{ICFDTC}{informed consent date character form}
  \item{DOBDT}{date of birth}
  \item{AGE}{subject baseline age}
  \item{SEX}{subject sex}
  \item{RACE}{subject race}
  \item{ETHNIC}{subject ethnicity}
}
}
\usage{
DM
}
\description{
Randomly generated demographic data domain
}
\keyword{datasets}
