% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.R
\docType{data}
\name{PC}
\alias{PC}
\title{PC}
\format{
## `PC`
A data frame with 420 rows and 19 variables:
\describe{
  \item{STUDYID}{study label}
  \item{SITEID}{site code}
  \item{USUBJID}{unique subject ID}
  \item{PCCAT}{domain category}
  \item{PCTEST}{analyte category}
  \item{VISIT}{visit label}
  \item{VISCRFN}{visit numeric code}
  \item{PCTPT}{timepoint label}
  \item{PCTPTN}{timepoint numeric code}
  \item{PCSTAT}{completion status}
  \item{PCDT}{observation date}
  \item{PCTM}{observation time}
  \item{PCTMC}{observation time character form}
  \item{PCDTC}{observation date and time}
  \item{PCORRES}{original pharmacokinetic observation}
  \item{PCORRESU}{original pharmacokinetic observation unit label}
  \item{PCSTRESN}{standardized pharmacokinetic observation}
  \item{PCSTRESU}{standardized pharmacokinetic observation unit label}
  \item{PCLLOQ}{standardized pharmacokinetic observation lower limit of quantification}
}
}
\usage{
PC
}
\description{
Randomly generated pharmacokinetic observation domain
}
\keyword{datasets}
