% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_assembly.R
\name{pk_build}
\alias{pk_build}
\title{Build a PK(PD) dataset}
\usage{
pk_build(
  ex,
  pc = NA,
  pd = NA,
  sl.cov = NA,
  tv.cov = NA,
  other = NA,
  time.units = "days",
  cycle.length = NA,
  na = -999,
  time.rnd = F,
  amt.rnd = F,
  dv.rnd = F,
  cov.rnd = F,
  impute = NA,
  BDV = F,
  DDV = F,
  PDV = F,
  sparse = 3,
  tv.cov.fill = "downup",
  col.order = NA
)
}
\arguments{
\item{ex}{dose event dataframe}

\item{pc}{pc event dataframe}

\item{pd}{pd event dataframe}

\item{sl.cov}{subject-level covarite dataframe}

\item{tv.cov}{time-varying covariate dataframe}

\item{other}{other event dataframe}

\item{time.units}{units for time attributes}

\item{cycle.length}{cycle length}

\item{na}{value for missing numeric items}

\item{time.rnd}{time attribute rounding parameter}

\item{amt.rnd}{amount attribute rounding parameter}

\item{dv.rnd}{dependent variable attribute rounding parameter}

\item{cov.rnd}{covariate attribute rounding parameter}

\item{impute}{imputation method}

\item{BDV}{baseline pd attribute}

\item{DDV}{change from baseline pd attribute}

\item{PDV}{percent change from baseline pd attribute}

\item{sparse}{threshold for sparse sampling}

\item{tv.cov.fill}{time-varying covariate fill direction}

\item{col.order}{column order}
}
\value{
PK(PD) dataset
}
\description{
Build a PK(PD) dataset
}
