% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_assembly.R
\name{pk_combine}
\alias{pk_combine}
\title{combine study level datasets to form population dataset}
\usage{
pk_combine(df1, df2, demo.map = T, na = -999)
}
\arguments{
\item{df1}{original PK(PD) dataset}

\item{df2}{additional PK(PD) dataset}

\item{demo.map}{toggle pre-set numeric values for SEX, RACE, and ETHNIC demographic variables}

\item{na}{value for missing numeric items}
}
\value{
population PK(PD) dataset
}
\description{
combine study level datasets to form population dataset
}
