% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_list_export.R
\name{variable_list_export}
\alias{variable_list_export}
\title{Export a standard variable list to support the pk_define() function.}
\usage{
variable_list_export(file)
}
\arguments{
\item{file}{desired filepath for the variable list}
}
\value{
standard variable list as a .csv file
}
\description{
Variable list outputted as a .csv file
The user should add additional definitions to the file for custom covariates.
}
\examples{
\dontshow{if (exists("vl_path")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Write variable list template to server with variable_list_export()
vl_path ##User designated filepath for variable list "C:/.../variable.list.csv"
if (file.exists(vl_path)) {
  variable_list_export(vl_path)
}
\dontshow{\}) # examplesIf}
}
