% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_assembly.R
\name{version_log}
\alias{version_log}
\title{Create and maintain a dataset version log}
\usage{
version_log(file, orig = F, outdir = NULL, prevdata = NULL, template, comp_var)
}
\arguments{
\item{file}{filepath of new dataset}

\item{orig}{original dataset flag}

\item{outdir}{output directory, defaults to dataset directory}

\item{prevdata}{comparison dataset filepath}

\item{template}{template .docx filepath}

\item{comp_var}{grouping variables for comparison}
}
\value{
version log as a .docx file
}
\description{
Version log is outputted as a .docx file
Document tracks changes in subject count, record count, new variables, and changing variables
User comments in the word document are preserved between versions
}
