% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VL.R
\docType{data}
\name{VL}
\alias{VL}
\title{VL}
\format{
## `VL`
A data frame with 56 rows and 5 variables:
\describe{
  \item{Variable}{Column or variable name}
  \item{Categorization}{Column or variable category}
  \item{Description}{Column or variable description}
  \item{Units}{Column or variable units (fixed)}
  \item{Comment}{NA by default}
}
}
\usage{
VL
}
\description{
Variable list with apmx variables and definitions
}
\keyword{datasets}
