% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_assembly.R
\name{pk_define}
\alias{pk_define}
\title{Create definition file from published dataset}
\usage{
pk_define(file, project, variable.list, template)
}
\arguments{
\item{file}{filepath of PK(PD) dataset}

\item{project}{project name}

\item{variable.list}{filepath of variable definition csv dataset}

\item{template}{filepath of docx definition file template}
}
\value{
NA
}
\description{
Create definition file from published dataset
}
