% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_assembly.R
\name{pk_define}
\alias{pk_define}
\title{Create definition file from published dataset}
\usage{
pk_define(
  file,
  project,
  variable.list,
  template,
  font = "Times New Roman",
  size = 9,
  na = -999
)
}
\arguments{
\item{file}{filepath of PK(PD) dataset}

\item{project}{project name}

\item{variable.list}{filepath of variable definition csv dataset}

\item{template}{filepath of docx definition file template}

\item{font}{font for table contents}

\item{size}{font size for table contents}

\item{na}{value used for missing or na numeric covariates}
}
\value{
NA
}
\description{
Definition file table can be read into a template word document (.docx) or blank document if desired.
Definitions are sourced from a variable list stored separately on your server.
Please refer to apmx::variable_list_export() for a standard copy of the variable list.
}
