% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_assembly.R
\name{variable_list_export}
\alias{variable_list_export}
\title{Export a standard variable list to support the pk_define() function.}
\usage{
variable_list_export(file)
}
\arguments{
\item{file}{desired filepath for the variable list}
}
\value{
standard variable list as a .csv file
}
\description{
Variable list outputted as a .csv file
The user should add additional definitions to the file for custom covariates.
}
